#3074-refactored notifications to separate router
