+++ src/backend/src/controllers/notifications.controllers.ts
@@ -11,4 +11,34 @@ export default class NotificationsController {
       next(error);
     }
   }
+
+  static async getUserUnreadNotifications(req: Request, res: Response, next: NextFunction) {
+    try {
+      const { organization, currentUser } = req;
+
+      const unreadNotifications = await NotificationsService.getUserUnreadNotifications(
+        currentUser.userId,
+        organization.organizationId
+      );
+      res.status(200).json(unreadNotifications);
+    } catch (error: unknown) {
+      next(error);
+    }
+  }
+
+  static async removeUserNotification(req: Request, res: Response, next: NextFunction) {
+    try {
+      const { notificationId } = req.body;
+      const { organization, currentUser } = req;
+
+      const unreadNotifications = await NotificationsService.removeUserNotification(
+        currentUser.userId,
+        notificationId,
+        organization.organizationId
+      );
+      res.status(200).json(unreadNotifications);
+    } catch (error: unknown) {
+      next(error);
+    }
+  }
 }
