+++ src/backend/src/routes/users.routes.ts
@@ -54,8 +54,6 @@ userRouter.post(
   validateInputs,
   UsersController.getManyUserTasks
 );
-userRouter.get('/notifications/current-user', UsersController.getUserUnreadNotifications);
 userRouter.get('/announcements/current-user', UsersController.getUserUnreadAnnouncements);
-userRouter.post('/notifications/remove', nonEmptyString(body('notificationId')), UsersController.removeUserNotification);
 
 export default userRouter;
