+++ src/backend/src/services/notifications.services.ts
@@ -197,6 +197,52 @@ export default class NotificationsService {
   }
 
   /**
+   * Gets all of a user's unread notifications
+   * @param userId id of user to get unread notifications from
+   * @param organization the user's orgainzation
+   * @returns the unread notifications of the user
+   */
+  static async getUserUnreadNotifications(userId: string, organizationId: string) {
+    const unreadNotifications = await prisma.notification.findMany({
+      where: {
+        users: {
+          some: { userId }
+        }
+      },
+      ...getNotificationQueryArgs(organizationId)
+    });
+
+    if (!unreadNotifications) throw new HttpException(404, 'User Unread Notifications Not Found');
+
+    return unreadNotifications.map(notificationTransformer);
+  }
+
+  /**
+   * Removes a notification from the user's unread notifications
+   * @param userId id of the current user
+   * @param notificationId id of the notification to remove
+   * @param organization the user's organization
+   * @returns the user's updated unread notifications
+   */
+  static async removeUserNotification(userId: string, notificationId: string, organizationId: string) {
+    const updatedUser = await prisma.user.update({
+      where: { userId },
+      data: {
+        unreadNotifications: {
+          disconnect: {
+            notificationId
+          }
+        }
+      },
+      include: { unreadNotifications: getNotificationQueryArgs(organizationId) }
+    });
+
+    if (!updatedUser) throw new HttpException(404, `Failed to remove notication: ${notificationId}`);
+
+    return updatedUser.unreadNotifications.map(notificationTransformer);
+  }
+
+  /**
    * Creates and sends a notification to all users with the given userIds
    * @param text writing in the notification
    * @param iconName icon that appears in the notification
