+++ src/backend/tests/unmocked/users.test.ts
@@ -3,7 +3,6 @@ import { createTestOrganization, createTestTask, createTestUser, resetUsers } fr
 import { batmanAppAdmin } from '../test-data/users.test-data';
 import UsersService from '../../src/services/users.services';
 import { NotFoundException } from '../../src/utils/errors.utils';
-import NotificationsService from '../../src/services/notifications.services';
 import AnnouncementService from '../../src/services/announcement.service';
 
 describe('User Tests', () => {
@@ -48,43 +47,6 @@ describe('User Tests', () => {
       const userTasks = await UsersService.getManyUserTasks([testBatman.userId, testBatman.userId], organization);
 
       expect(userTasks).toStrictEqual([batmanTask, batmanTask]);
-    });
-  });
-
-  describe('Get Notifications', () => {
-    it('Succeeds and gets user notifications', async () => {
-      const testBatman = await createTestUser(batmanAppAdmin, orgId);
-      await NotificationsService.sendNotifcationToUsers('test1', 'test1', [testBatman.userId], orgId);
-      await NotificationsService.sendNotifcationToUsers('test2', 'test2', [testBatman.userId], orgId);
-
-      const notifications = await UsersService.getUserUnreadNotifications(testBatman.userId, organization);
-
-      expect(notifications).toHaveLength(2);
-      expect(notifications[0].text).toBe('test1');
-      expect(notifications[1].text).toBe('test2');
-    });
-  });
-
-  describe('Remove Notifications', () => {
-    it('Succeeds and removes user notification', async () => {
-      const testBatman = await createTestUser(batmanAppAdmin, orgId);
-      await NotificationsService.sendNotifcationToUsers('test1', 'test1', [testBatman.userId], orgId);
-      await NotificationsService.sendNotifcationToUsers('test2', 'test2', [testBatman.userId], orgId);
-
-      const notifications = await UsersService.getUserUnreadNotifications(testBatman.userId, organization);
-
-      expect(notifications).toHaveLength(2);
-      expect(notifications[0].text).toBe('test1');
-      expect(notifications[1].text).toBe('test2');
-
-      const updatedNotifications = await UsersService.removeUserNotification(
-        testBatman.userId,
-        notifications[0].notificationId,
-        organization
-      );
-
-      expect(updatedNotifications).toHaveLength(1);
-      expect(updatedNotifications[0].text).toBe('test2');
     });
   });
