+++ src/frontend/src/utils/urls.ts
@@ -26,8 +26,6 @@ const userScheduleSettings = (id: string) => `${usersById(id)}/schedule-settings
 const userScheduleSettingsSet = () => `${users()}/schedule-settings/set`;
 const userTasks = (id: string) => `${usersById(id)}/tasks`;
 const manyUserTasks = () => `${users()}/tasks/get-many`;
-const userNotifications = () => `${users()}/notifications/current-user`;
-const userRemoveNotifications = () => `${users()}/notifications/remove`;
 
 /**************** Projects Endpoints ****************/
 const projects = () => `${API_URL}/projects`;
@@ -196,6 +194,11 @@ const faqCreate = () => `${recruitment()}/faq/create`;
 const faqEdit = (id: string) => `${recruitment()}/faq/${id}/edit`;
 const faqDelete = (id: string) => `${recruitment()}/faq/${id}/delete`;
 
+/************** Notification Endpoints ***************/
+const notifications = () => `${API_URL}/notifications`;
+const notificationsCurrentUser = () => `${notifications()}/current-user`;
+const notificationsRemoveCurrentUser = () => `${notificationsCurrentUser()}/remove`;
+
 /**************** Other Endpoints ****************/
 const version = () => `https://api.github.com/repos/Northeastern-Electric-Racing/FinishLine/releases/latest`;
 
@@ -214,8 +217,6 @@ export const apiUrls = {
   userScheduleSettingsSet,
   userTasks,
   manyUserTasks,
-  userNotifications,
-  userRemoveNotifications,
 
   projects,
   allProjects,
@@ -354,6 +355,10 @@ export const apiUrls = {
   faqCreate,
   faqEdit,
   faqDelete,
+
+  notifications,
+  notificationsCurrentUser,
+  notificationsRemoveCurrentUser,
 
   version
 };
